Code has been optimized and readme information has been updated
